login layout added, packages reordered, dummy textview, activity, test renamed
