+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="greenfox.com.zerdareader">
+    package="com.greenfox.fuchsit.zerdareader">
 
     <application
         android:allowBackup="true"
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".LoginActivity">
+        <activity android:name="com.greenfox.fuchsit.zerdareader.activity.HelloWorldActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
