+++ app/src/main/java/com/greenfox/fuchsit/zerdareader/activity/HelloWorldActivity.java
@@ -1,13 +1,14 @@
-package greenfox.com.zerdareader;
+package com.greenfox.fuchsit.zerdareader.activity;
 
-import android.R.layout;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
-public class LoginActivity extends AppCompatActivity {
+import com.greenfox.fuchsit.zerdareader.R;
+
+public class HelloWorldActivity extends AppCompatActivity {
 
     Button button;
     TextView textView;
@@ -15,7 +16,7 @@ public class LoginActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.login);
+        setContentView(R.layout.hello_world);
 
         button = (Button) findViewById(R.id.button);
         button.setOnClickListener(new View.OnClickListener() {
