--- app/src/main/res/layout/login_new.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/login"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="greenfox.com.zerdareader.LoginActivity">
-
-    <LinearLayout
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <TextView
-            android:text="TextView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/loginTitle"
-            android:textSize="30sp"
-            tools:text="Please login"
-            android:textAlignment="center"
-            android:layout_marginBottom="20dp"/>
-
-        <LinearLayout
-            android:orientation="vertical"
-            android:layout_width="218dp"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:background="@drawable/customborder">
-
-            <EditText
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:inputType="textPersonName"
-                android:hint="username"
-                android:ems="10"
-                android:id="@+id/userName"
-                android:textAlignment="center"
-                android:textSize="24sp" />
-
-            <EditText
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:inputType="textPassword"
-                android:hint="password"
-                android:ems="10"
-                android:textAlignment="center"
-                android:id="@+id/password"
-                android:textSize="22sp" />
-
-            <Button
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick = "login"
-                android:id="@+id/loginButton"
-                tools:text="LOGIN" />
-        </LinearLayout>
-    </LinearLayout>
-
-</RelativeLayout>
