+++ app/src/test/java/com/greenfox/fuchsit/zerdareader/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package greenfox.com.zerdareader;
+package com.greenfox.fuchsit.zerdareader;
 
 import org.junit.Test;
